requirements.txt
flask
requests
google-cloud-firestore
google-cloud-pubsub
numpy
gunicorn
wtforms
flask-restful